refactor: util -> infra 패키지 이동, `NumberGenerator`테스트
