+++ src/test/java/infrastructure/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package util;
+package infrastructure;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
